warehouse/mwaa/requirements.txt
--constraint https://raw.githubusercontent.com/apache/airflow/constraints-3.0.6/constraints-3.12.txt

astronomer-cosmos>=1.10,<1.11